Move useTeamForm custom hook inside TeamsForm component
